feat: allow unencoded `userinfo` as per SIP002 (#260)
* feat: allow unencoded `userinfo` as per SIP002

* Move shadowsocks config tests to `shadowsocks_test.go` and clean up.

* Review comment.

* Add test case for invalid cipher info (e.g. missin colon).

* Remove unnecessary prefix in test.
