+++ x/config/config_test.go
@@ -15,7 +15,6 @@
 package config
 
 import (
-	"encoding/base64"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -55,69 +54,9 @@ func TestSanitizeConfig(t *testing.T) {
 	require.Error(t, err)
 }
 
-func TestShowsocksLagacyBase64URL(t *testing.T) {
-	encoded := base64.URLEncoding.WithPadding(base64.NoPadding).EncodeToString([]byte("aes-256-gcm:1234567@example.com:1234?prefix=HTTP%2F1.1%20"))
-	urls, err := parseConfig("ss://" + string(encoded) + "#outline-123")
-	require.NoError(t, err)
-	require.Equal(t, 1, len(urls))
-	config, err := parseShadowsocksLegacyBase64URL(urls[0])
-	require.Equal(t, "example.com:1234", config.serverAddress)
-	require.Equal(t, "HTTP/1.1 ", string(config.prefix))
-	require.NoError(t, err)
-}
-
-func TestParseShadowsocksURL(t *testing.T) {
-	encoded := base64.URLEncoding.WithPadding(base64.NoPadding).EncodeToString([]byte("aes-256-gcm:1234567@example.com:1234?prefix=HTTP%2F1.1%20"))
-	urls, err := parseConfig("ss://" + string(encoded) + "#outline-123")
-	require.NoError(t, err)
-	require.Equal(t, 1, len(urls))
-	config, err := parseShadowsocksURL(urls[0])
-	require.Equal(t, "example.com:1234", config.serverAddress)
-	require.Equal(t, "HTTP/1.1 ", string(config.prefix))
-	require.NoError(t, err)
-
-	encoded = base64.URLEncoding.WithPadding(base64.NoPadding).EncodeToString([]byte("aes-256-gcm:1234567"))
-	urls, err = parseConfig("ss://" + string(encoded) + "@example.com:1234?prefix=HTTP%2F1.1%20" + "#outline-123")
-	require.NoError(t, err)
-	require.Equal(t, 1, len(urls))
-	config, err = parseShadowsocksURL(urls[0])
-	require.Equal(t, "example.com:1234", config.serverAddress)
-	require.Equal(t, "HTTP/1.1 ", string(config.prefix))
-	require.NoError(t, err)
-}
-
 func TestSocks5URLSanitization(t *testing.T) {
 	configString := "socks5://myuser:mypassword@192.168.1.100:1080"
 	sanitizedConfig, err := SanitizeConfig(configString)
 	require.NoError(t, err)
 	require.Equal(t, "socks5://REDACTED@192.168.1.100:1080", sanitizedConfig)
-}
-
-func TestParseShadowsocksSIP002URLUnsuccessful(t *testing.T) {
-	encoded := base64.URLEncoding.WithPadding(base64.NoPadding).EncodeToString([]byte("aes-256-gcm:1234567@example.com:1234?prefix=HTTP%2F1.1%20"))
-	urls, err := parseConfig("ss://" + string(encoded) + "#outline-123")
-	require.NoError(t, err)
-	require.Equal(t, 1, len(urls))
-	_, err = parseShadowsocksSIP002URL(urls[0])
-	require.Error(t, err)
-}
-
-func TestParseShadowsocksSIP002URLUnsupportedCypher(t *testing.T) {
-	configString := "ss://Y2hhY2hhMjAtaWV0Zi1wb2x5MTMwnTpLeTUyN2duU3FEVFB3R0JpQ1RxUnlT@example.com:1234?prefix=HTTP%2F1.1%20"
-	urls, err := parseConfig(configString)
-	require.NoError(t, err)
-	require.Equal(t, 1, len(urls))
-	_, err = parseShadowsocksSIP002URL(urls[0])
-	require.Error(t, err)
-}
-
-func TestParseShadowsocksSIP002URLSuccessful(t *testing.T) {
-	configString := "ss://Y2hhY2hhMjAtaWV0Zi1wb2x5MTMwNTpLeTUyN2duU3FEVFB3R0JpQ1RxUnlT@example.com:1234?prefix=HTTP%2F1.1%20"
-	urls, err := parseConfig(configString)
-	require.NoError(t, err)
-	require.Equal(t, 1, len(urls))
-	config, err := parseShadowsocksSIP002URL(urls[0])
-	require.NoError(t, err)
-	require.Equal(t, "example.com:1234", config.serverAddress)
-	require.Equal(t, "HTTP/1.1 ", string(config.prefix))
 }
